requirements/test.txt
# Test dependencies go here.
-r base.txt

django_mock_queries==1.0.5
pytest
pytest-cov==2.4.0
coverage==4.2
pytest-django
pytest-mock>=1,<2
tox
vcrpy>=1.0.0
ipdb
